requirements.txt
asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.13.0
gunicorn==20.1.0
waitress==2.1.2
whitenoise==6.2.0
Pillow==8.1.2
SQLAlchemy==1.3.22
pytz==2021.1
python-decouple==3.6
django-heroku==0.3.1
djangorestframework==3.13.1
